+++ src/main/python/plotlyst/view/widget/manuscript/sprint.py
@@ -1,0 +1,135 @@
+"""
+Plotlyst
+Copyright (C) 2021-2025  Zsolt Kovari
+
+This file is part of Plotlyst.
+
+Plotlyst is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+Plotlyst is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+"""
+
+import datetime
+from typing import Optional
+
+from PyQt6.QtCore import QUrl
+from PyQt6.QtMultimedia import QSoundEffect
+from PyQt6.QtWidgets import QWidget
+from qthandy import retain_when_hidden, transparent
+from qthandy.filter import OpacityEventFilter
+from qtmenu import MenuWidget
+
+from plotlyst.common import RELAXED_WHITE_COLOR
+from plotlyst.core.sprint import TimerModel
+from plotlyst.resources import resource_registry
+from plotlyst.view.common import ButtonPressResizeEventFilter
+from plotlyst.view.generated.sprint_widget_ui import Ui_SprintWidget
+from plotlyst.view.generated.timer_setup_widget_ui import Ui_TimerSetupWidget
+from plotlyst.view.icons import IconRegistry
+
+
+class TimerSetupWidget(QWidget, Ui_TimerSetupWidget):
+    def __init__(self, parent=None):
+        super(TimerSetupWidget, self).__init__(parent)
+        self.setupUi(self)
+
+    def value(self) -> int:
+        return self.sbTimer.value() * 60
+
+
+class SprintWidget(QWidget, Ui_SprintWidget):
+    def __init__(self, parent=None):
+        super().__init__(parent)
+        self.setupUi(self)
+        self._model = None
+        self._compact: bool = False
+        self.setModel(TimerModel())
+
+        self._toggleState(False)
+        transparent(self.time)
+        transparent(self.btnPause)
+        transparent(self.btnReset)
+
+        self.btnTimer.setIcon(IconRegistry.timer_icon())
+        self.btnPause.installEventFilter(OpacityEventFilter(self.btnPause, leaveOpacity=0.7))
+        self.btnPause.installEventFilter(ButtonPressResizeEventFilter(self.btnPause))
+        self.btnReset.setIcon(IconRegistry.restore_alert_icon('#ED6868'))
+        self.btnReset.installEventFilter(OpacityEventFilter(self.btnReset, leaveOpacity=0.7))
+        self.btnReset.installEventFilter(ButtonPressResizeEventFilter(self.btnReset))
+
+        self._timer_setup = TimerSetupWidget()
+        self._menu = MenuWidget(self.btnTimer)
+        self._menu.addWidget(self._timer_setup)
+
+        self._timer_setup.btnStart.clicked.connect(self.start)
+        self.btnPause.clicked.connect(self._pauseStartTimer)
+        self.btnReset.clicked.connect(self._reset)
+
+        self._effect: Optional[QSoundEffect] = None
+
+    def model(self) -> TimerModel:
+        return self._model
+
+    def setModel(self, model: TimerModel):
+        self._model = model
+        self._model.valueChanged.connect(self._updateTimer)
+        self._model.finished.connect(self._finished)
+        self._toggleState(self._model.isActive())
+
+    def setCompactMode(self, compact: bool):
+        self._compact = compact
+        self._toggleState(self.model().isActive())
+        self.time.setStyleSheet(f'border: 0px; color: {RELAXED_WHITE_COLOR}; background-color: rgba(0,0,0,0);')
+
+    def start(self):
+        self._toggleState(True)
+        self._model.start(self._timer_setup.value())
+        self._updateTimer()
+        self._menu.close()
+
+    def _toggleState(self, running: bool):
+        self.time.setVisible(running)
+        if running:
+            self.btnPause.setChecked(True)
+            self.btnPause.setIcon(IconRegistry.pause_icon(color='grey'))
+        if self._compact:
+            self.btnTimer.setHidden(running)
+            retain_when_hidden(self.btnPause)
+            retain_when_hidden(self.btnReset)
+            self.btnPause.setHidden(True)
+            self.btnReset.setHidden(True)
+        else:
+            self.btnPause.setVisible(running)
+            self.btnReset.setVisible(running)
+
+    def _updateTimer(self):
+        mins, secs = self._model.remainingTime()
+        time = datetime.time(minute=mins, second=secs)
+        self.time.setTime(time)
+
+    def _pauseStartTimer(self, played: bool):
+        self.model().toggle()
+        if played:
+            self.btnPause.setIcon(IconRegistry.pause_icon(color='grey'))
+        else:
+            self.btnPause.setIcon(IconRegistry.play_icon())
+
+    def _reset(self):
+        self.model().stop()
+        self._toggleState(False)
+
+    def _finished(self):
+        if self._effect is None:
+            self._effect = QSoundEffect()
+            self._effect.setSource(QUrl.fromLocalFile(resource_registry.cork))
+            self._effect.setVolume(0.3)
+        self._effect.play()
